refactor(user): using paths aliases from tsconfig
