+++ nodemon.json
@@ -4,7 +4,7 @@
   ],
   "ext": "js,ts,json",
   "ignore": [
-    "src/**/*.test.ts"
+    "tests/**/*"
   ],
-  "exec": "ts-node ./src/api/main.ts"
+  "exec": "ts-node -r tsconfig-paths/register ./src/api/main.ts"
 }
