+++ package.json
@@ -22,7 +22,7 @@
     "@commitlint/cli": "19.5.0",
     "@commitlint/config-conventional": "19.5.0",
     "@eslint/js": "9.11.1",
-    "@tsconfig/node20": "20.1.4",
+    "@tsconfig/node22": "22.0.0",
     "@types/bcrypt": "5.0.2",
     "@types/express": "5.0.0",
     "@types/jest": "29.5.13",
@@ -39,6 +39,7 @@
     "prisma": "6.0.1",
     "ts-jest": "29.2.5",
     "ts-node": "10.9.2",
+    "tsconfig-paths": "4.2.0",
     "typescript": "5.5.4",
     "typescript-eslint": "8.7.0"
   },
