+++ src/Contexts/user/domain/dtos/update-user-persistence.dto.ts
@@ -1,4 +1,4 @@
-import { UserPrimitive } from '../primtives/user.primitive'
+import { UserPrimitive } from '../primitives/user.primitive'
 
 export type UpdateUserPersistenceDTO = Partial<
   Omit<UserPrimitive, 'id' | 'password' | 'ctime'>
