+++ src/Contexts/user/domain/dtos/user-response.dto.ts
@@ -1,3 +1,3 @@
-import { UserPrimitive } from '../primtives/user.primitive'
+import { UserPrimitive } from '../primitives/user.primitive'
 
 export type UserResponseDTO = Omit<UserPrimitive, 'password'>
