+++ src/Contexts/user/domain/entities/user.entity.ts
@@ -3,7 +3,7 @@ import { UserEmail } from '../value-objects/UserEmail'
 import { UserId } from '../value-objects/UserId'
 import { UserName } from '../value-objects/UserName'
 import { UserPassword } from '../value-objects/UserPassword'
-import { UserPrimitive } from '../primtives/user.primitive'
+import { UserPrimitive } from '../primitives/user.primitive'
 import { UserTime } from '../value-objects/UserTime'
 import { CreateUserDTO } from '../dtos/create-user.dto'
 import { UpdateUserPersistenceDTO } from '../dtos/update-user-persistence.dto'
