+++ src/Contexts/user/domain/value-objects/UserEmail.ts
@@ -1,3 +1,3 @@
-import { StringValueObject } from '../../../shared/domain/value-objects/string.value-object'
+import { StringValueObject } from '@contexts/shared/domain/value-objects/string.value-object'
 
 export class UserEmail extends StringValueObject {}
