+++ src/Contexts/user/domain/value-objects/UserId.ts
@@ -1,3 +1,3 @@
-import { UuidValueObject } from '../../../shared/domain/value-objects/uuid.value-object'
+import { UuidValueObject } from '@contexts/shared/domain/value-objects/uuid.value-object'
 
 export class UserId extends UuidValueObject {}
