+++ src/Contexts/user/domain/value-objects/UserPassword.ts
@@ -1,5 +1,5 @@
-import { StringValueObject } from '../../../shared/domain/value-objects/string.value-object'
-import { PasswordUtils } from '../../../shared/utils/password.utils'
+import { StringValueObject } from '@contexts/shared/domain/value-objects/string.value-object'
+import { PasswordUtils } from '@contexts/shared/utils/password.utils'
 
 export class UserPassword extends StringValueObject {
   static async create(plainPassword: string): Promise<UserPassword> {
