+++ src/Contexts/user/domain/value-objects/UserTime.ts
@@ -1,3 +1,3 @@
-import { DateValueObject } from '../../../shared/domain/value-objects/date.value-object'
+import { DateValueObject } from '@contexts/shared/domain/value-objects/date.value-object'
 
 export class UserTime extends DateValueObject {}
