+++ src/api/config/di-container.ts
@@ -1,12 +1,12 @@
 import { PrismaClient } from '@prisma/client'
-import { UserPrismaRepository } from '../../Contexts/user/infrastructure/repositories/user-prisma.repository'
+import { UserPrismaRepository } from '@contexts/user/infrastructure/repositories/user-prisma.repository'
 
-import { UsersController } from '../controllers/users.controller'
-import { CreateUserService } from '../../Contexts/user/application/create-user.service'
-import { DeleteUserService } from '../../Contexts/user/application/delete-user.service'
-import { GetAllUsersService } from '../../Contexts/user/application/get-all-users.service'
-import { GetUserByIdService } from '../../Contexts/user/application/get-user-by-id.service'
-import { UpdateUserService } from '../../Contexts/user/application/update-user.service'
+import { UsersController } from '@api/controllers/users.controller'
+import { CreateUserService } from '@contexts/user/application/create-user.service'
+import { DeleteUserService } from '@contexts/user/application/delete-user.service'
+import { GetAllUsersService } from '@contexts/user/application/get-all-users.service'
+import { GetUserByIdService } from '@contexts/user/application/get-user-by-id.service'
+import { UpdateUserService } from '@contexts/user/application/update-user.service'
 
 const prismaClient = new PrismaClient()
 const userPrismaRepository = new UserPrismaRepository(prismaClient)
