+++ src/api/controllers/users.controller.ts
@@ -1,15 +1,15 @@
 import { NextFunction, Request, Response } from 'express'
-import { CreateUserService } from '../../Contexts/user/application/create-user.service'
-import { DeleteUserService } from '../../Contexts/user/application/delete-user.service'
-import { GetAllUsersService } from '../../Contexts/user/application/get-all-users.service'
-import { GetUserByIdService } from '../../Contexts/user/application/get-user-by-id.service'
-import { UpdateUserService } from '../../Contexts/user/application/update-user.service'
-import { UserValidator } from '../../Contexts/user/domain/validation/user.validator'
-import { BaseError } from '../../Contexts/shared/domain/errors/base.error'
-import { UnexpectedError } from '../../Contexts/shared/domain/errors/unexpected.error'
-import { UserResponseDTO } from '../../Contexts/user/domain/dtos/user-response.dto'
-import { UpdateUserDTO } from '../../Contexts/user/domain/dtos/update-user.dto'
-import { CreateUserDTO } from '../../Contexts/user/domain/dtos/create-user.dto'
+import { CreateUserService } from '@contexts/user/application/create-user.service'
+import { DeleteUserService } from '@contexts/user/application/delete-user.service'
+import { GetAllUsersService } from '@contexts/user/application/get-all-users.service'
+import { GetUserByIdService } from '@contexts/user/application/get-user-by-id.service'
+import { UpdateUserService } from '@contexts/user/application/update-user.service'
+import { UserValidator } from '@contexts/user/domain/validation/user.validator'
+import { BaseError } from '@contexts/shared/domain/errors/base.error'
+import { UnexpectedError } from '@contexts/shared/domain/errors/unexpected.error'
+import { UserResponseDTO } from '@contexts/user/domain/dtos/user-response.dto'
+import { UpdateUserDTO } from '@contexts/user/domain/dtos/update-user.dto'
+import { CreateUserDTO } from '@contexts/user/domain/dtos/create-user.dto'
 
 export class UsersController {
   constructor(
