+++ tests/__mocks__/prisma/users/updateUser.input.ts
@@ -1,4 +1,4 @@
-import { UpdateUserDTO } from '../../../../src/Contexts/user/domain/dtos/update-user.dto'
+import { UpdateUserDTO } from '@contexts/user/domain/dtos/update-user.dto'
 
 export const updateUserInput: UpdateUserDTO = {
   name: 'johndoe',
